refactor(compare): share driver helpers
Extract the still-identical compare driver helper layer so direct compare and compare-outputs reuse the same shared structs, counters, path bucketing, and dependency identity helpers while leaving divergent license-delta logic local.
