+++ src/lib.rs
@@ -37,6 +37,8 @@ pub mod cache;
 pub mod cli;
 pub(crate) mod compare;
 #[doc(hidden)]
+pub mod compare_driver_shared;
+#[doc(hidden)]
 pub mod compare_normalization;
 pub mod copyright;
 pub mod finder;
